Also present test matrix rows as flat rows
It turns out that using hashes as matrix rows makes it impossible to
exclude combinations such as Ubuntu 20.04 + Puppet 5 (for which no
packages exist).

The chosen solution is to add additional definitions which can be
mapped.

In the case of setfiles, it's preferred to have a matrix based
on beaker_setfile_names and then use beaker_setfile_values with the name
as a key. This gives the actual setfile.

There is no method data to build the collection. This is rather trivial
and probably keeps the actual use easier.

All old structures are still present for compatibility.