Add support for Ubuntu 22.04
Puppet packages are not yet available for this new version of this
Debian derivative, so tests are likely to fail.  This is however
preferable to not running the tests on this version of buntu and missing
failures when adding compatibility with Ubuntu 22.04.

Support of Ubuntu 22.04 for Puppet can be tracked here:
https://tickets.puppetlabs.com/browse/PA-4233

While here, complete metadata about Ubuntu releases expiration dates.