Add --beaker-facter option
This changes things so you can run a test matrix by providing an
additional fact.

The use case is that you have a test suite for your software and want to
support multiple versions of that software. For example, in
puppet-pulpcore there are versions 3.21, 3.22 and 3.28 supported at the
same time. Rather than designing the test suite to first test x, then y
and then z while cleaning up in between, this allows providing the
version as a fact (so setup_acceptance_node.rb can set up repos). You
then expand the test matrix. It also includes the version number so it's
easy to see which version of the software is being tested. That it can
use parallelism for this is just a side effect, but a nice one.

To make it more flexible the config for GHA is changed by providing
puppet_beaker_test_matrix. This only provides a name and a set of
environment variables. This means other CI systems could also use this.

Only a single fact can be provided. Generating the whole matrix was too
complex and not needed for my use case. Because it's only providing a
name and a set of env vars, it could be extended in the future.