Skip EOL operating systems in GHA acceptance tests
With CentOS Linux 7 and CentOS Stream 8 going EOL soon, these will start
to fail because mirrors will no longer be available. This creates an
annotation and skips the OS release.

It introduces parameters to pretend it's a different date to keep the
test suite consistent and avoid breaking it automatically over time.